backend/requirements.txt
aioredis==1.2.0
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==19.1.1
Automat==0.7.0
certifi==2018.11.29
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
daphne==2.2.5
Django==2.1.5
django-anymail==5.0
django-cors-headers==2.4.0
django-environ==0.4.5
djangorestframework==3.9.1
environ==1.0
hiredis==1.0.0
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
PyHamcrest==1.9.0
pytz==2018.9
requests==2.21.0
six==1.12.0
tinys3==0.1.12
Twisted==18.9.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
